chore: Add iOS assets and update Firebase Hosting configuration
This commit adds the necessary iOS assets to the project, including the "Contents.json" file for the "Assets.xcassets" directory and the "Contents.json" file for the "Icon.appiconset" directory. Additionally, it updates the Firebase Hosting source directory to "build/web" in the firebase.json file. These changes are made to ensure proper configuration and deployment of the project.